+++ typescript/tests/fwlist/random-stats.test.ts
@@ -25,8 +25,8 @@ describe.for(
     
     test("zero probabilities", () => {
       test_binomial(method, new FWList(
-        [0, "no"],
         [1, "yes"],
+        [0, "no"],
       ));
     });
     
@@ -39,6 +39,39 @@ describe.for(
       test_binomial(method, new FWList(
         [1000, "sup"],
         [1, "ayo"],
+      ));
+    });
+  });
+
+  describe("fractional weights", () =>
+  {
+    test("equal probabilities", () => {
+      test_binomial(method, new FWList(
+        [0.5, "left"],
+        [0.5, "right"],
+      ));
+    });
+    
+    // test("usual probabilities", () => {
+    //   test_binomial(method, fwl().normalised());
+    // });
+    
+    test("zero probabilities", () => {
+      test_binomial(method, new FWList(
+        [0.1, "yes"],
+        [0, "no"],
+      ));
+    });
+    
+    test("extreme probabilities", () => {
+      test_binomial(method, new FWList(
+        [10, "sup"],
+        [0.1, "ayo"],
+      ));
+      
+      test_binomial(method, new FWList(
+        [100, "sup"],
+        [0.1, "ayo"],
       ));
     });
   });